아이템 대거 추가
아이템 대거 추가

+++ ProjectBS/Assets/_BsScripts/Item/ItemData/ItemData.cs
@@ -10,7 +10,7 @@ public abstract class ItemData : ScriptableObject
 
     public int ID => _id;
     public string Name => _name;
-    public float Value => _value;
+    public int Value => _value;
     public GameObject ItemPrefab => _prefab;
     // public itemtype TypeOfItem => _itemType;
     [Multiline]
@@ -18,7 +18,7 @@ public abstract class ItemData : ScriptableObject
 
     [SerializeField] private int _id;               
     [SerializeField] private string _name;
-    [SerializeField] private float _value;
+    [SerializeField] private int _value;
     [SerializeField] private GameObject _prefab;
     //[SerializeField] private itemtype _itemType;
     // Start is called before the first frame update
